refactor: BoolQuery 생성 로직을 Cond 클래스 안으로 이동
