+++ src/test/java/com/bookspot/book/infra/search/cond/BookCategoryCondTest.java
@@ -1,6 +1,12 @@
 package com.bookspot.book.infra.search.cond;
 
+import com.bookspot.category.application.BookCategoryDto;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -26,6 +32,22 @@ class BookCategoryCondTest {
         BookCategoryCond categoryCond = BookCategoryCond.top(400, "자연과학");
         assertEquals(BookCategoryCond.TOP_FILED, categoryCond.getCategoryField());
         assertEquals("400.자연과학", categoryCond.getCategoryValue());
+    }
+
+
+    @ParameterizedTest(name = "{0}")
+    @MethodSource("args")
+    void 카테고리_이름_생성(String expected, int id, String name) {
+        assertEquals(expected, BookCategoryCond.buildValue(id, name));
+    }
+
+    static Stream<Arguments> args() {
+        return Stream.of(
+                Arguments.of("000.총류", 0, "총류"),
+                Arguments.of("010.도서학, 서지학", 10, "도서학, 서지학"),
+                Arguments.of("011.저작", 11, "저작"),
+                Arguments.of("376.중등교육", 376, "중등교육")
+        );
     }
 
 }
